:truck: Rename commons.rs to core.rs and update imports
Renamed cli/src/command/commons.rs to cli/src/command/core.rs and updated all module imports and references from 'commons' to 'core' across the command modules. This improves clarity and consistency in module naming.
