+++ cli/src/command.rs
@@ -4,9 +4,9 @@ pub mod bugreport;
 mod chmod;
 mod chown;
 mod chunk;
-mod commons;
 pub mod complete;
 pub(crate) mod concat;
+mod core;
 pub mod create;
 mod delete;
 pub mod diff;
