+++ cli/src/command/core.rs
