+++ cli/src/command/migrate.rs
@@ -2,7 +2,7 @@ use crate::{
     cli::{PasswordArgs, SolidEntriesTransformStrategy, SolidEntriesTransformStrategyArgs},
     command::{
         ask_password,
-        commons::{
+        core::{
             collect_split_archives, run_transform_entry, TransformStrategyKeepSolid,
             TransformStrategyUnSolid,
         },
